backend-app/requirements.txt
boto3==1.9.106
botocore==1.12.106
cent==3.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.4.1
django-js-asset==1.2.2
django-mptt==0.9.1
django-storages==1.7.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docutils==0.14
filetype==1.0.5
freeze==1.0.10
idna==2.8
jmespath==0.9.4
mysqlclient==1.4.2.post1
Pillow==5.4.1
PyJWT==1.7.1
pymediainfo==3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
s3transfer==0.2.0
short-url==1.2.2
six==1.12.0
urllib3==1.24.1
